Modified download to zip files and send to frontend, modified upload to work with zip files and directories

+++ backend/app.py
@@ -1,12 +1,13 @@
 import boto3
 import logging
 import os
+import shutil
 from botocore.exceptions import ClientError
-from flask import Flask, jsonify, request
-from zipfile import ZipFile
+from flask import Flask, jsonify, request, send_file
 
 main_bucket = "bucket-for-testing-boto3"
 zip_temp = "zip_temp"
+download_temp = "Sanctum_Images"
 
 app = Flask(__name__)
 
@@ -17,26 +18,31 @@ app = Flask(__name__)
 def test_function():
     return jsonify(list_existing_buckets())
 
-# Downloads the specified image to the user's current directory
-# Input: file: file to download
-# Output: Bool
+# Downloads an array of images from s3 bucket and sends them to frontend
+# Input: files: Array of files to download
+# Output: Zipped up directory containing downloaded files, 
+#         Number of files that were not downloaded
 @app.route('/download', methods=['POST'])
-def download_file():
+def download_files():
     data = request.get_json()
     
-    if 'file' in data:
+    if 'files' in data:
         try:
-            success = download_file(data['file'], main_bucket)
-            return jsonify(success)
+            files_not_found = download_files(data['files'], main_bucket)
+            if os.path.exists(download_temp + ".zip"):
+                send_file(download_temp + ".zip", as_attachment=True)
+                return jsonify({"Files not found": files_not_found})
+            else:
+                return jsonify({"error": "requested file(s) not found"}), 500
         except Exception as e:
             return jsonify({"error": str(e)}), 500
     else:
         return jsonify({"error": "Incorrect input format"}), 400
     
-# Uploads an array of files to the database
+# Uploads an array of files to the database, works with zip files and directories
 # Input: files: Array of files
 # Output: Bool
-@app.route('/upload_files', methods=['POST'])
+@app.route('/upload', methods=['POST'])
 def upload_files():
     data = request.get_json()
 
@@ -68,13 +74,16 @@ def upload_files(file_arr, bucket):
         for file in file_arr:
             if file.endswith(".zip"):
                 unzip_files(file)
+            elif os.path.isdir(file):
+                upload_dir(file, bucket)
             else:
                 s3_client.upload_file(file, bucket, file)
-        for dir_, _, files in os.walk(zip_temp):
-            for file in files:
-                path = os.path.join(dir_,file)
-                s3_client.upload_file(path, bucket, file)
-                os.remove(path)
+        if (os.path.exists(zip_temp)):
+            for dir_, _, files in os.walk(zip_temp): # walk avoids the issue of uploading directories
+                for file in files:
+                    path = os.path.join(dir_,file)
+                    s3_client.upload_file(path, bucket, file)
+            shutil.rmtree(zip_temp)
     except ClientError as e:
         logging.error(e)
         return False
@@ -86,21 +95,34 @@ def upload_dir(directory, bucket):
     print(cmd)
     os.system(cmd)
 
-# Downloads a file into the user's current directory
-def download_file(file_name, bucket):
+# Downloads a file into the temporary folder download_temp, then zips it up and deletes the original
+def download_files(file_arr, bucket):
     boto3.setup_default_session(profile_name='dev')
     s3 = boto3.client('s3')
-    try:
-        s3.download_file(bucket, file_name, file_name)
-    except ClientError as e:
-        logging.error(e)
-        return False
-    return True
+
+    if os.path.exists(download_temp): # remove old download file if it exists
+        shutil.rmtree(download_temp)
+    os.mkdir(download_temp)
+
+    if os.path.exists(download_temp + ".zip"): # remove old zip file if it exists
+        os.remove(download_temp + ".zip")
+
+    not_found = 0
+
+    for file in file_arr:
+        try:
+            s3.download_file(bucket, file, os.path.join(download_temp, file))
+        except ClientError as e:
+            logging.error(e)
+            not_found += 1
+    if len([i for i in os.listdir(download_temp)]) > 0:
+        shutil.make_archive(download_temp, "zip", download_temp)
+    shutil.rmtree(download_temp)
+    return not_found
 
 # internal function that handles zip files. Uses temporary folder zip_temp
 def unzip_files(file_name):
-    with ZipFile(file_name, 'r') as zip:
-        zip.extractall(zip_temp)
+    shutil.unpack_archive(file_name, zip_temp)
 
 if __name__ == '__main__':
     app.run()
